Cleaning up Code
- Moved Perform Photometry Functions to own modules
- Changed name of Perform_Photometry function to Perform_PSF_Photometry
- Added GUI Options for Warner, Buchhiem and Boyde Methods
- Added Some Documentation
